dbr quant overhead[5/x]: remove unnecessary asserts
Summary:

Removes asserts which are duplicate (the same condition is checked
when calculating the hook type, so there is no need to check it again).
For the assert in `validate_is_at_last_seen_idx`, rewrites it to
raise an Error instead to ensure it does not get stripped in
production environments.

Test plan:

```
python test/test_quantization.py TestQuantizeDBR
```

ghstack-source-id: 7f6f150c19c4524a27c27f02f83a2a3cf5baa93a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68370